@@ -13,18 +13,25 @@
     "@next/env": "^14.2.7",
     "@radix-ui/react-label": "^2.1.0",
     "@vercel/postgres": "^0.9.0",
+    "accept-language": "^3.0.20",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "cors": "^2.8.5",
     "cross-fetch": "^4.0.0",
     "formik": "^2.4.6",
+    "i18next": "^23.15.1",
+    "i18next-browser-languagedetector": "^8.0.0",
+    "i18next-resources-to-backend": "^1.2.1",
     "js-cookie": "^3.0.5",
     "jsonwebtoken": "^9.0.2",
     "lucide-react": "^0.427.0",
     "next": "14.2.5",
     "next-auth": "^5.0.0-beta.20",
-    "react": "^18",
-    "react-dom": "^18",
+    "next-i18next": "^15.3.1",
+    "react": "^18.3.1",
+    "react-cookie": "^7.2.0",
+    "react-dom": "^18.3.1",
+    "react-i18next": "^15.0.1",
     "tailwind-merge": "^2.5.2",
     "tailwindcss-animate": "^1.0.7",
     "vercel": "^32.3.0",
