@@ -2,57 +2,56 @@
 
 import React from 'react';
 import { useMutation, gql } from '@apollo/client';
-import { CreateEventSchema } from './../lib/validations'
+import { CreateEventSchema } from '../../lib/validations';
 import { Formik, Form, Field, ErrorMessage } from 'formik';
-
+import { useTranslation } from './../../../i18n/client'; // Importación para traducción
 
 const CREATE_EVENT_MUTATION = gql`
   mutation CreateEvent($name: String!, $description: String!, $location: String!, $date: String!, $maxCapacity: Int!) {
     createEvent(name: $name, description: $description, location: $location, date: $date, maxCapacity: $maxCapacity) {
+      id
+      name
+      description
+      location
+      date
+      maxCapacity
+      organizer {
         id
         name
-        description
-        location
-        date
-        maxCapacity
-        organizer {
-            id
-            name
-            email
-            role
-        }
+        email
+        role
+      }
     }
   }
 `;
 
-export default function CreateEvent() {
-  const [formData, setFormData] = React.useState(new FormData());
+export default function CreateEvent({params}) {
+  const { t } = useTranslation(params.lang, 'create-event'); // Hook de traducción
 
   const [createEvent, { data, loading, error }] = useMutation(CREATE_EVENT_MUTATION, {
     onCompleted: (data) => {
       console.log('Event created:', data.createEvent);
       window.location.href = '/dashboard';
-
     },
   });
 
-  const handleSubmit = async (values,{setSubmitting}) => {
-    const {name, description, location , date, maxCapacity } = values;
+  const handleSubmit = async (values, { setSubmitting }) => {
+    const { name, description, location, date, maxCapacity } = values;
     
     try {
       await createEvent({
         variables: {
-          name: name,
-          description: description,
-          location: location,
-          date: date,
+          name,
+          description,
+          location,
+          date,
           maxCapacity: parseInt(maxCapacity, 10),
         },
       });
     } catch (error) {
       console.error('Error creating event:', error);
     } finally {
-      setSubmitting(false); // Indicar que el envío ha terminado
+      setSubmitting(false);
     }
   };
 
@@ -68,7 +67,7 @@ export default function CreateEvent() {
     <div className="min-h-screen bg-gray-100 py-12 px-4 sm:px-6 lg:px-8">
       <div className="max-w-lg mx-auto p-4">
         <div className="mt-6 bg-white shadow overflow-hidden sm:rounded-lg">
-          <h1 className="text-2xl font-bold m-4">Crear Evento</h1>
+          <h1 className="text-2xl font-bold m-4">{t('title')}</h1>
           <Formik
             initialValues={initialValues}
             validationSchema={CreateEventSchema}
@@ -78,7 +77,7 @@ export default function CreateEvent() {
               <Form className="space-y-2 m-4">
                 <div>
                   <label htmlFor="name" className="block text-sm font-medium text-gray-700">
-                    Nombre de Evento
+                    {t('name')}
                   </label>
                   <Field
                     type="text"
@@ -90,7 +89,7 @@ export default function CreateEvent() {
                 </div>
                 <div>
                   <label htmlFor="description" className="block text-sm font-medium text-gray-700">
-                    Descripción
+                    {t('description')}
                   </label>
                   <Field
                     as="textarea"
@@ -102,7 +101,7 @@ export default function CreateEvent() {
                 </div>
                 <div>
                   <label htmlFor="location" className="block text-sm font-medium text-gray-700">
-                    Lugar
+                    {t('location')}
                   </label>
                   <Field
                     type="text"
@@ -114,7 +113,7 @@ export default function CreateEvent() {
                 </div>
                 <div>
                   <label htmlFor="date" className="block text-sm font-medium text-gray-700">
-                    Fecha
+                    {t('date')}
                   </label>
                   <Field
                     type="date"
@@ -126,7 +125,7 @@ export default function CreateEvent() {
                 </div>
                 <div>
                   <label htmlFor="maxCapacity" className="block text-sm font-medium text-gray-700">
-                    Capacidad Máxima
+                    {t('maxCapacity')}
                   </label>
                   <Field
                     type="number"
@@ -139,12 +138,12 @@ export default function CreateEvent() {
                 <button
                   type="submit"
                   className="mt-4 w-full bg-indigo-500 text-white py-2 rounded-md hover:bg-indigo-700"
-                  disabled={loading||isSubmitting}
+                  disabled={loading || isSubmitting}
                 >
-                  {loading ? 'Creando...' : 'Crear Evento'}
+                  {loading ? t('loading') : t('submit')}
                 </button>
-                {error && <p className="mt-2 text-red-500">Error: {error.message}</p>}
-                {data && <p className="mt-2 text-green-500">Evento creado exitosamente</p>}
+                {error && <p className="mt-2 text-red-500">{t('error')}: {error.message}</p>}
+                {data && <p className="mt-2 text-green-500">{t('success')}</p>}
               </Form>
             )}
           </Formik>
