@@ -1,8 +1,8 @@
-"use client"
+"use client";
 
 import { useQuery, gql } from '@apollo/client';
+import { useTranslation } from './../../../i18n/client';
 
-// Definimos la query para obtener el usuario autenticado y sus eventos aplicados
 const CURRENT_USER_APPLIED_EVENTS_QUERY = gql`
   query {
     user {
@@ -32,24 +32,23 @@ const CURRENT_USER_APPLIED_EVENTS_QUERY = gql`
   }
 `;
 
-
-export default function ProfilePage() {
-
+export default function ProfilePage({params}) {
+  const { t } = useTranslation(params.lang,'profile'); // Hook para traducción
   const { loading, error, data } = useQuery(CURRENT_USER_APPLIED_EVENTS_QUERY);
 
-  if (loading) return <p>Loading...</p>;
-  if (error) return <p>Error: {error.message}</p>;
+  if (loading) return <p>{t('loading')}</p>;
+  if (error) return <p>{t('error_loading')} {error.message}</p>;
 
-  const {user} = data;
+  const { user } = data;
 
   return (
     <div className="min-h-screen bg-gray-100 py-12 px-4 sm:px-6 lg:px-8">
       <div className="max-w-7xl mx-auto">
-        <h1 className="text-3xl font-extrabold text-gray-900">Perfil de {user.name}</h1>
-        
+        <h1 className="text-3xl font-extrabold text-gray-900">{t('profile_title', { name: user.name })}</h1>
+
         <div className="mt-6 bg-white shadow overflow-hidden sm:rounded-lg">
           <div className="px-4 py-5 sm:px-6">
-            <h3 className="text-lg font-medium text-gray-900">Eventos Creados</h3>
+            <h3 className="text-lg font-medium text-gray-900">{t('created_events')}</h3>
           </div>
           <div className="border-t border-gray-200">
             <ul role="list" className="divide-y divide-gray-200">
@@ -64,12 +63,12 @@ export default function ProfilePage() {
                   </div>
                   <div className="ml-4 flex-shrink-0">
                     <a href={`/events/${event.id}`} className="text-indigo-600 hover:text-indigo-500">
-                      Ver detalles
+                      {t('view_details')}
                     </a>
                   </div>
                   <div>
-                    <a href={`/profile/manage/${event.id}`} className="text-center ml-2 inline-flex  px-2 py-2 border border-transparent text-sm font-medium rounded-md text-white bg-indigo-600 hover:bg-indigo-700">
-                      Gestionar aplicaciones
+                    <a href={`/profile/manage/${event.id}`} className="text-center ml-2 inline-flex px-2 py-2 border border-transparent text-sm font-medium rounded-md text-white bg-indigo-600 hover:bg-indigo-700">
+                      {t('manage_applications')}
                     </a>
                   </div>
                 </li>
@@ -80,7 +79,7 @@ export default function ProfilePage() {
 
         <div className="mt-6 bg-white shadow overflow-hidden sm:rounded-lg">
           <div className="px-4 py-5 sm:px-6">
-            <h3 className="text-lg font-medium text-gray-900">Eventos Aplicados</h3>
+            <h3 className="text-lg font-medium text-gray-900">{t('applied_events')}</h3>
           </div>
           <div className="border-t border-gray-200">
             <ul role="list" className="divide-y divide-gray-200">
@@ -91,11 +90,11 @@ export default function ProfilePage() {
                       <p className="text-gray-900 font-medium">{application.event.name}</p>
                       <p className="text-gray-500">{new Date(parseInt(application.event.date)).toLocaleDateString()}</p>
                       <p className="text-gray-500">{application.event.location}</p>
-                      <p className="text-gray-500">Status: {application.status}</p>
+                      <p className="text-gray-500">{t('status')}: {application.status}</p>
                     </div>
                   </div>
                   <div className="ml-4 flex-shrink-0">
-                    <a href={`/events/${application.event.id}`} className="text-indigo-600 hover:text-indigo-500">Ver detalles</a>
+                    <a href={`/events/${application.event.id}`} className="text-indigo-600 hover:text-indigo-500">{t('view_details')}</a>
                   </div>
                 </li>
               ))}
